Add LICENSE file for Spotify Cookie Extractor For Spotube
This commit adds the GNU Affero General Public License (AGPL) v3 to the project. The license ensures that the modified source code of the Spotify Cookie Extractor For Spotube Chrome extension is made available to the community, especially in cases where the software is used on network servers. The license text includes the following information:

- Program name and description
- Copyright notice for 2024 Silvan Fülle
- Terms of the AGPL v3 license, as published by the Free Software Foundation

This license ensures the software remains free and open-source while protecting the rights and contributions of the developers.